requirements.txt
django==5.0.6
djangorestframework

# For testing
pytest
pytest-cov
pytest-django